+++ src/android/app/src/main/java/io/github/alexdenton/interviewd/template/detail/TemplateDetailActivity.kt
@@ -20,7 +20,7 @@ class TemplateDetailActivity : RxAwareActivity() {
 
         if (savedInstanceState == null) {
             vm.initWith(LazyKodein(appKodein))
-            vm.useId(receive() as Int)
+            vm.useId(receive() as Long)
 
             fragmentTransaction {
                 add(R.id.templateDetail_fragmentContainer, TemplateDetailFragment())
