+++ src/android/app/src/online/java/io/github/alexdenton/interviewd/retrofit/dto/TemplateDto.kt
@@ -3,5 +3,5 @@ package io.github.alexdenton.interviewd.retrofit.dto
 /**
  * Created by ryan on 8/14/17.
  */
-data class TemplateDto (val name: String, val questionIds: List<Int>, val id: Int = 0){
+data class TemplateDto (val name: String, val questionIds: List<Long>, val id: Long = 0){
 }
