+++ src/android/app/src/online/java/io/github/alexdenton/interviewd/retrofit/repositories/CandidateRetrofitRepository.kt
@@ -1,12 +1,11 @@
 package io.github.alexdenton.interviewd.retrofit.repositories
 
 import android.content.Context
-import io.github.alexdenton.interviewd.api.InterviewdApi
-import io.github.alexdenton.interviewd.retrofit.RetrofitFactory
-import io.github.alexdenton.interviewd.retrofit.dto.CandidateDto
 import io.github.alexdenton.interviewd.api.repositories.CandidateRepository
 import io.github.alexdenton.interviewd.entities.Candidate
 import io.github.alexdenton.interviewd.retrofit.InterviewdApiRetrofit
+import io.github.alexdenton.interviewd.retrofit.RetrofitFactory
+import io.github.alexdenton.interviewd.retrofit.dto.CandidateDto
 import io.reactivex.Single
 
 /**
@@ -16,7 +15,7 @@ class CandidateRetrofitRepository(val context: Context) : CandidateRepository {
 
     val client: InterviewdApiRetrofit = RetrofitFactory(context).create(RetrofitFactory.Mode.Local)
 
-    override fun getCandidate(id: Int): Single<Candidate>
+    override fun getCandidate(id: Long): Single<Candidate>
             = client.getCandidate(id)
             .map { it.toCandidate() }
 
@@ -32,7 +31,7 @@ class CandidateRetrofitRepository(val context: Context) : CandidateRepository {
             = client.patchCandidate(candidate.id, candidate.toDto())
             .map { it.toCandidate() }
 
-    override fun deleteCandidate(id: Int): Single<Candidate>
+    override fun deleteCandidate(id: Long): Single<Candidate>
             = client.deleteCandidate(id)
             .map { it.toCandidate() }
